removed unnecessary codes and moved functions to logic
