+++ app/static/js/createEvents.js
@@ -551,60 +551,4 @@ $(".startDatePicker, .endDatePicker").change(function () {
   });
 
   setCharacterLimit($("#inputCharacters"), "#remainingCharacters"); 
-
-  const eventId = $("#eventId").val();
-  if (eventId) {
-    const selectedCohorts = selectedCohorts(eventId);
-    
-    $('input[name="cohorts[]"]').each(function() {
-      const cohortYear = $(this).val();
-      if (selectedCohorts.includes(cohortYear)) {
-        $(this).prop('checked', true);
-      }
-    });
-    
-    $('input[name="cohorts[]"]').on('change', function() {
-      const selectedCohorts = $('input[name="cohorts[]"]:checked')
-        .map(function() { return $(this).val(); })
-        .get();
-      saveSelectedCohorts(eventId, selectedCohorts);
-    });
-    
-    getCohortStatus(eventId);
-  }
 });
-
-
-function getCohortStatus(eventId) {
-  return $.ajax({
-    type: "GET",
-    url: `/event/${eventId}/cohortStatus`,
-    success: function(response) {
-      displayCohortStatus(response.invitedCohorts);
-    }
-  });
-}
-
-function displayCohortStatus(invitedCohorts) {
-  $(".cohortStatus").remove(); 
-  
-  invitedCohorts.forEach(cohort => {
-    const invitedDate = new Date(cohort.invited_at).toLocaleDateString();
-    const cohortElement = $(`#cohort-${cohort.year}`);
-    
-    if (cohortElement.length) {
-      const statusHtml = `
-        <div class="cohortStatus mt-1 ml-4">
-          <small class="text-success">
-            <i class="fas fa-check-circle"></i>
-            Invited on ${invitedDate}
-          </small>
-        </div>
-      `;
-      cohortElement.closest('.form-check').append(statusHtml);
-      cohortElement.prop('checked', true);
-      cohortElement.prop('disabled', true);
-      cohortElement.closest('.form-check-label').addClass('text-muted');
-    }
-  });
-}
